net: dsa: sja1105: set the pvid on the DSA links as 4095
Since commit ed040abca4c1 ("net: dsa: sja1105: use 4095 as the private
VLAN for untagged traffic"), this driver uses a reserved value as pvid
for the host port (DSA CPU port). Control packets which are sent as
untagged get classified to this VLAN, and all ports are members of it
(this is to be expected for control packets).

Manage all DSA links (host ports of downstream switches and cascade
ports) in the same way and allow control packets to egress everywhere.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>