net: dsa: sja1105: enable address learning on DSA links
There are 3 types of packets traversing DSA links which need to be
considered:
(a) packets destined for a remote switch for bridging.
(b) packets reaching the CPU for local termination through a bridge.
(c) packets reaching the CPU for local standalone port termination.

Since the sja1105 driver has unconditionally enabled address learning on
the CPU port so far without many issues, it seems fairly low risk to do
so on DSA links as well.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>