net: dsa: tag_sja1105: suppress the "Couldn't decode source port" message
H topologies like this one still have a problem:

         eth0                                                     eth1
          |                                                        |
       CPU port                                                CPU port
          |                        DSA link                        |
 sw0p0  sw0p1  sw0p2  sw0p3  sw0p4 -------- sw1p4  sw1p3  sw1p2  sw1p1  sw1p0
   |             |      |                            |      |             |
 user          user   user                         user   user          user
 port          port   port                         port   port          port

Basically any packet sent by the eth0 DSA master can be flooded on the
interconnecting DSA link sw0p4 <-> sw1p4 and it will be received by the
eth1 DSA master too. Basically we are talking to ourselves.

When the bridge is VLAN-unaware, the packet is encoded using a tag_8021q
TX VLAN, dsa_8021q_rcv() rightfully cannot decode it and complains.
We can fix this problem by not allowing a TX VLAN to egress towards the
CPU port, just ingress from the CPU port.

However this can not be done with VLAN-aware bridges. There we have no
such "smart" rule based on VLAN membership, because the VLAN in the
packet is the VLAN known by the bridge, and both CPU ports need to be in
the broadcast domain of all bridge VLANs.

So the only thing we can do is to let DSA silently drop them.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>