net: dsa: sja1105: suppress TX VLANs from looping back in "H" topologies
H topologies like this one still have a problem:

         eth0                                                     eth1
          |                                                        |
       CPU port                                                CPU port
          |                        DSA link                        |
 sw0p0  sw0p1  sw0p2  sw0p3  sw0p4 -------- sw1p4  sw1p3  sw1p2  sw1p1  sw1p0
   |             |      |                            |      |             |
 user          user   user                         user   user          user
 port          port   port                         port   port          port

Basically any packet sent by the eth0 DSA master can be flooded on the
interconnecting DSA link sw0p4 <-> sw1p4 and it will be received by the
eth1 DSA master too. Basically we are talking to ourselves.

But since this packet is encoded using a tag_8021q TX VLAN,
dsa_8021q_rcv() rightfully cannot decode it and complains.

The proper solution to this problem is to not allow a TX VLAN to egress
towards the CPU port, just ingress from the CPU port.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>